Refactor Base class and change package location to "commands.updateactions"
