+++ commercetools-internal-processors/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -8,3 +8,4 @@ io.sphere.sdk.annotations.processors.UpdateCommandEndpointAnnotationProcessor
 io.sphere.sdk.annotations.processors.DeleteCommandEndpointAnnotationProcessor
 io.sphere.sdk.annotations.processors.HasQueryModelAnnotationProcessor
 io.sphere.sdk.annotations.processors.HasQueryModelImplementationAnnotationProcessor
+io.sphere.sdk.annotations.processors.HasUpdateActionsProcessor
